whisper/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
asttokens==3.0.0
attrs==25.3.0
audio2numpy==0.1.2
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.13.3
botocore==1.37.23
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
datasets==2.14.5
decorator==5.2.1
deprecation==2.1.0
dill==0.3.7
executing==2.2.0
ffmpeg==1.4
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2023.6.0
gdown==4.7.1
gitdb==4.0.12
GitPython==3.1.42
h5py==3.13.0
huggingface-hub==0.23.5
humanfriendly==10.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.15
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
onnx==1.16.2
onnxruntime==1.21.0
onnxruntime-qnn==1.21.0
openai-whisper==20231117
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
prettytable==3.11.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==3.20.3
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
Pygments==2.19.1
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qai-hub==0.26.0
qai-hub-models==0.25.5
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
samplerate==0.2.1
schema==0.7.5
scipy==1.15.2
semver==3.0.4
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tiktoken==0.9.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
# AnythingLLM dependencies
httpx==0.28.1